# secp256k1: Overhaul compact signatures.