--- a/dcrec/secp256k1/signature_bench_test.go
+++ b/dcrec/secp256k1/signature_bench_test.go
@@ -106,7 +106,7 @@ func BenchmarkSignCompact(b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		_, _ = SignCompact(privKey, msgHash, true)
+		_ = SignCompact(privKey, msgHash, true)
 	}
 }