**This requires #2094**.

This overhauls compact signatures to significantly optimize them and to use the recently-added mod n scalar type along with field vals.  It also significantly improves the comments and readability while overhauling the related code.

The existing approach to create a compact signature was extremely inefficient since it involved attempting to recover all of the possible public key candidates from the signature and comparing them against the real public key in order to determine which candidate was the correct one.

This takes a much more efficient approach that consists of modifying the signing code to produce and return a "pubkey recovery code" that uniquely identifies which of the possible candidates is correct since the signing code already has the necessary information.  It adds virtually zero overhead to normal signatures, but has a significant impact on compact signatures.

It also gains speed advantages by making use of several of the recently-added optimized field val methods such as y coordinate decompression and determining if a field val is >= the field prime minus the group order.

The changes have been split up into multiple commits to help ease the review process.

The following benchmarks show a before and after comparison of producing regular signatures, producing compact signatures, and recovering the public key from the compact signatures:

```
benchmark                 old ns/op    new ns/op    delta
-----------------------------------------------------------
BenchmarkSign             58268        58478         +0.36%
BenchmarkSignCompact      551174       58171        -89.45%
BenchmarkRecoverCompact   212780       206042       -3.17%

benchmark                 old allocs   new allocs   delta
-----------------------------------------------------------
BenchmarkSign             36           36            +0.00%
BenchmarkSignCompact      166          39           -76.51%
BenchmarkRecoverCompact   51           36           -29.41%

benchmark                 old bytes    new bytes    delta
-----------------------------------------------------------
BenchmarkSign             1776         1776          +0.00%
BenchmarkSignCompact      7815         1920         -75.43%
BenchmarkRecoverCompact   2610         1745         -33.14%
```